# openApiDoc/openapi.yaml
openapi: 3.0.3
info:
  title: Car Sales
  description: Car Sales API
  contact:
    email: sbc.mfreitas@outlook.com
  version: 1.0.0
servers:
- url: http://localhost:8080/v1
...
paths:
  /v1/car:
    get:
      tags:
        - Car
      summary: Return a list of cars
      description: Return a list of cars
      operationId: getCarsUsingGET
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CarResponse'
    put:
      tags:
        - Car
      summary: Update an existing car
      description: Update an existing car
      operationId: updateCarUsingPUT
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarResponse'
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: Internal server Error
    post:
      tags:
        - Car
      summary: Create new Car
      description: Create new Car
      operationId: createCarUsingPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarResponse'
        '400':
          description: Bad request
        '404':
          description: Not Found
        '500':
          description: Internal server Error
  '/v1/car/{id}':
    get:
      tags:
        - Car
      summary: Find car by ID
      description: Find the car by ID
      operationId: findCarByIdUsingGET
      parameters:
        - name: id
          in: path
          description: ID of car to return
          required: true
          style: simple
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarResponse'
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: Internal server Error
    delete:
      tags:
        - Car
      summary: Delete a car
      description: Delete a car
      operationId: deleteCarUsingDELETE
      parameters:
        - name: id
          in: path
          description: Car id to delete
          required: true
          style: simple
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
        '404':
          description: Not found
        '500':
          description: Internal server Error
...
components:
  schemas:
    BrandResponse:
      title: BrandResponse
      required:
        - id
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    CarRequest:
      title: CarRequest
      required:
        - brandId
        - id
        - optionIds
        - value
      type: object
      properties:
        brandId:
          type: integer
          format: int64
        id:
          type: integer
          format: int64
        optionIds:
          type: array
          items:
            type: integer
            format: int64
        value:
          type: number
          format: bigdecimal
    CarResponse:
      title: CarResponse
      required:
        - id
        - options
        - value
      type: object
      properties:
        brand:
          $ref: '#/components/schemas/BrandResponse'
        id:
          type: integer
          format: int64
        options:
          type: array
          items:
            $ref: '#/components/schemas/OptionResponse'
        value:
          type: number
          format: bigdecimal
...
    OptionResponse:
      title: OptionResponse
      required:
        - id
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
...